using cached campus pdfs

+++ lib/features/home/presentation/widgets/settings_page.dart
@@ -31,7 +31,7 @@ class _SettingsPageState extends State<SettingsPage> {
       body: SingleChildScrollView(
         physics: const ClampingScrollPhysics(),
         primary: true,
-        padding: const EdgeInsets.symmetric(horizontal: 20, vertical: 20),
+        padding: const EdgeInsets.symmetric(horizontal: 10, vertical: 10),
         child: Column(
           mainAxisAlignment: MainAxisAlignment.start,
           crossAxisAlignment: CrossAxisAlignment.center,
@@ -70,9 +70,7 @@ class _SettingsPageState extends State<SettingsPage> {
                         activeTrackColor: Theme.of(context).primaryColor,
                         inactiveThumbColor:
                             Theme.of(context).colorScheme.onPrimary,
-                        inactiveTrackColor: Theme.of(context)
-                            .colorScheme
-                            .scrim,
+                        inactiveTrackColor: Theme.of(context).colorScheme.scrim,
                         trackOutlineColor:
                             const WidgetStatePropertyAll(Colors.transparent),
                       ),
@@ -101,9 +99,48 @@ class _SettingsPageState extends State<SettingsPage> {
                 text: "Sign Out",
                 icon: Icons.logout_rounded,
                 onTap: () async {
-                  final storage = const FlutterSecureStorage();
-                  await storage.delete(key: "user");
-                  GoRouter.of(context).goNamed(UhlLinkRoutesNames.chooseAuth);
+                  await showDialog(
+                    context: context,
+                    builder: (context) {
+                      return AlertDialog(
+                        title: Text("Sign Out",
+                            style: Theme.of(context).textTheme.bodyMedium),
+                        content: Text(
+                            "Are you sure you want to sign out of Vertex?",
+                            style: Theme.of(context).textTheme.labelSmall),
+                        backgroundColor: Theme.of(context).cardColor,
+                        shape: RoundedRectangleBorder(
+                          borderRadius: BorderRadius.circular(10),
+                        ),
+                        actionsPadding:
+                            const EdgeInsets.only(right: 10, bottom: 10),
+                        actions: [
+                          TextButton(
+                            onPressed: () => Navigator.of(context).pop(),
+                            child: Text("Cancel",
+                                style: Theme.of(context).textTheme.labelSmall),
+                          ),
+                          TextButton(
+                            onPressed: () async {
+                              final storage = const FlutterSecureStorage();
+                              await storage.delete(key: "user");
+                              GoRouter.of(context)
+                                  .goNamed(UhlLinkRoutesNames.chooseAuth);
+                            },
+                            child: Text("Sign Out",
+                                style: Theme.of(context)
+                                    .textTheme
+                                    .labelSmall!
+                                    .copyWith(
+                                      color:
+                                          Theme.of(context).colorScheme.onError,
+                                      fontWeight: FontWeight.w700,
+                                    )),
+                          ),
+                        ],
+                      );
+                    },
+                  );
                 }),
           ],
         ),
